fix: moving sync function out of async
